Clean up PKGBUILD
Rather than linking to raw files, this will link to the latest stable
tarball of the source.

Also includes some style changes, preferring a programatic approach to
dealing with the package's name and version where possible.

Freenode/#archlinux users Earnestly, polyzen, and alad helped with this.

This commit changes the name back to simply `pia-tools`.